+++ yash-syntax/CHANGELOG.md
@@ -13,6 +13,20 @@ Terminology: A _public dependency_ is one that’s exposed through this crate’
 public API (e.g., re-exported types).
 A _private dependency_ is used internally and not visible to downstream users.
 
+## [0.22.1] - Unreleased
+
+### Added
+
+- `parser::lex::Lexer::mode` and `parser::lex::Lexer::set_mode` for querying and
+  updating the parsing mode (`yash_env::parser::Mode`) of a lexer. The parser
+  and lexer will consult the mode to decide which syntax to accept; setting the
+  `portable` mode currently has no effect on parsing yet.
+
+### Changed
+
+- Public dependency versions:
+    - yash-env 0.15.0 → 0.15.3
+
 ## [0.22.0] - 2026-06-11
 
 ### Changed
@@ -670,6 +684,7 @@ command.
 - Functionalities to parse POSIX shell scripts
 - Alias substitution support
 
+[0.22.1]: https://github.com/magicant/yash-rs/releases/tag/yash-syntax-0.22.1
 [0.22.0]: https://github.com/magicant/yash-rs/releases/tag/yash-syntax-0.22.0
 [0.21.0]: https://github.com/magicant/yash-rs/releases/tag/yash-syntax-0.21.0
 [0.20.0]: https://github.com/magicant/yash-rs/releases/tag/yash-syntax-0.20.0